+++ CHANGES.md
@@ -1,3 +1,8 @@
+### unreleased
+
+- Add `Duration.to_(seconds|minutes|hours)` ([gr-im](https://github.com/gr-im))
+- Add an implementation for `Lunar` ([gr-im](https://github.com/gr-im))
+
 ### v1.0.0
 
 - First release of `lunar` exposing _among other things_ `Time`, `Date`, `Datetime`, `Timezone`, `Zoned_datetime` and `Range` (by [xvw](https://github.com/xvw), [gr-im](https://github.com/gr-im)).